# backend/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Platform Import Export API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/v1/billing/status":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "facturation"
                                                                      ],
                                                             "summary":  "Get Billing Status",
                                                             "operationId":  "get_billing_status_api_v1_billing_status_get",
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "Successful Response",
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "type":  "object",
                                                                                                                                                   "title":  "Response Get Billing Status Api V1 Billing Status Get"
                                                                                                                                               }
                                                                                                                                }
                                                                                                       }
                                                                                       }
                                                                           },
                                                             "security":  [
                                                                              {
                                                                                  "HTTPBearer":  [

                                                                                                 ]
                                                                              }
                                                                          ]
                                                         }
                                             },
                  "/api/v1/billing/create-payment-intent":  {
                                                                "post":  {
                                                                             "tags":  [
                                                                                          "facturation"
                                                                                      ],
                                                                             "summary":  "Create Payment",
                                                                             "operationId":  "create_payment_api_v1_billing_create_payment_intent_post",
                                                                             "security":  [
                                                                                              {
                                                                                                  "HTTPBearer":  [

                                                                                                                 ]
                                                                                              }
                                                                                          ],
                                                                             "parameters":  [
                                                                                                {
                                                                                                    "name":  "type_paiement",
                                                                                                    "in":  "query",
                                                                                                    "required":  true,
                                                                                                    "schema":  {
                                                                                                                   "type":  "string",
                                                                                                                   "title":  "Type Paiement"
                                                                                                               }
                                                                                                },
                                                                                                {
                                                                                                    "name":  "success_url",
                                                                                                    "in":  "query",
                                                                                                    "required":  false,
                                                                                                    "schema":  {
                                                                                                                   "type":  "string",
                                                                                                                   "default":  "http://localhost:3000/success",
                                                                                                                   "title":  "Success Url"
                                                                                                               }
                                                                                                },
                                                                                                {
                                                                                                    "name":  "cancel_url",
                                                                                                    "in":  "query",
                                                                                                    "required":  false,
                                                                                                    "schema":  {
                                                                                                                   "type":  "string",
                                                                                                                   "default":  "http://localhost:3000/cancel",
                                                                                                                   "title":  "Cancel Url"
                                                                                                               }
                                                                                                }
                                                                                            ],
                                                                             "responses":  {
                                                                                               "200":  {
                                                                                                           "description":  "Successful Response",
                                                                                                           "content":  {
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {

                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                       }
                                                                                                       },
                                                                                               "422":  {
                                                                                                           "description":  "Validation Error",
                                                                                                           "content":  {
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                       }
                                                                                                       }
                                                                                           }
                                                                         }
                                                            },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   },
                       "securitySchemes":  {
                                               "HTTPBearer":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer"
                                                              }
                                           }
                   }
}